# swagger.yaml
openapi: 3.0.3
info:
  title: SHTT API - Quản lý Sở hữu Trí tuệ
  description: |
    API quản lý Sở hữu Trí tuệ gồm: Bản quyền tác giả, Nhãn hiệu, Sáng chế,
    Kiểu dáng công nghiệp, Chỉ dẫn địa lý, Làng nghề truyền thống.

    ## Xác thực
    Sử dụng **JWT Bearer Token**. Đăng nhập qua `/api/auth/login` để lấy `accessToken`,
    sau đó nhấn nút **Authorize** và nhập `Bearer <token>`.

    ## Phân quyền
    | Nhóm endpoint | Quyền yêu cầu |
    |---|---|
    | GET /search, GET /, GET /:id | Công khai |
    | GET /search-all | Admin hoặc Staff |
    | POST / PUT / DELETE (IP resources) | Admin hoặc Staff |
    | Mọi thao tác /api/users | Chỉ Admin |

    ## Upload ảnh
    Các endpoint POST/PUT sử dụng `multipart/form-data`.
    Trường `owner`, `authors`, `managementOrg` truyền dưới dạng **JSON string** trong form.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Access Token lấy từ POST /api/auth/login
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Thao tác thành công
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Lỗi server
        error:
          type: string
          example: Error detail

    # ── Stakeholder ─────────────────────────────
...
    UserObject:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1714100000000
        username:
          type: string
          example: admin
        fullName:
          type: string
          example: Quản trị viên
        role:
          type: string
          enum: [admin, staff]
          example: admin

    # ── Copyright ───────────────────────────────
...
  responses:
    Unauthorized:
      description: Chưa xác thực — thiếu hoặc sai token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Unauthorized
    NotFound:
      description: Không tìm thấy tài nguyên
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Không tìm thấy
    ServerError:
      description: Lỗi server nội bộ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Lỗi server
            error: Error detail

# ─────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────
paths:
...
  /api/users:
    post:
      tags: [Users]
      summary: Tạo tài khoản mới
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, fullName, password, role]
              properties:
                username:
                  type: string
                  example: staff_user
                fullName:
                  type: string
                  example: Nguyễn Văn A
                password:
                  type: string
                  example: SecurePass@2024
                role:
                  type: string
                  enum: [admin, staff]
                  example: staff
      responses:
        '201':
          description: Tạo thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Tạo tài khoản thành công
                  data:
                    $ref: '#/components/schemas/UserObject'
        '400':
          description: Thiếu thông tin bắt buộc
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Username đã tồn tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: ID người dùng (timestamp milliseconds)
        example: 1714100000000
    get:
      tags: [Users]
      summary: Xem chi tiết tài khoản
      operationId: getUserById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/UserObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags: [Users]
      summary: Cập nhật tài khoản
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                  example: Nguyễn Văn A - Đã cập nhật
                role:
                  type: string
                  enum: [admin, staff]
                password:
                  type: string
                  description: Bỏ trống nếu không đổi mật khẩu
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Cập nhật tài khoản thành công
                  data:
                    $ref: '#/components/schemas/UserObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Users]
      summary: Xóa tài khoản
      operationId: deleteUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Xóa thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # ════════════════════════════════
  # STAKEHOLDERS
  # ════════════════════════════════
...